+++ web/src/components/componentRegistry.ts
@@ -1,3 +1,4 @@
+import InputAmount from './InputAmount.vue';
 import Input from './Input.vue';
 import InputNumber from './InputNumber.vue';
 import InputPhone from './InputPhone.vue';
@@ -5,6 +6,7 @@ import type { ComponentName } from './Component';
 
 export const FORM_COMPONENT_MAP = {
   Input,
+  InputAmount,
   InputNumber,
   InputPhone,
 } as const satisfies Record<ComponentName, unknown>;